requirements.txt
Django==1.6.7
South==1.0
django-custom-user==0.5
djangorestframework==2.4.3
djangorestframework-camel-case==0.2.0
psycopg2==2.5.4
wsgiref==0.1.2
